requirements.txt
openai~=0.27.4
discord~=2.2.2
spacy~=3.5.2
levenshtein
attrs
